rocksdb: header file cleanup
